+++ src/server/controlchan/handlers/list.rs
@@ -15,7 +15,7 @@
 
 use super::handler::CommandContext;
 use crate::server::controlchan::handlers::ControlCommandHandler;
-use crate::server::controlchan::handlers::Command;
+use crate::server::controlchan::Command;
 use crate::server::error::FTPError;
 use crate::server::reply::{Reply, ReplyCode};
 use crate::storage;
