+++ src/server/controlchan/handlers/mod.rs
@@ -10,11 +10,10 @@ mod allo;
 mod auth;
 mod ccc;
 mod cdup;
-mod handler;
-mod command;
 mod cwd;
 mod dele;
 mod feat;
+mod handler;
 mod help;
 mod list;
 mod mdtm;
@@ -51,11 +50,10 @@ pub use allo::Allo;
 pub use auth::{Auth, AuthParam};
 pub use ccc::Ccc;
 pub use cdup::Cdup;
-pub(crate) use handler::{ControlCommandHandler, CommandContext};
-pub use command::Command;
 pub use cwd::Cwd;
 pub use dele::Dele;
 pub use feat::Feat;
+pub(crate) use handler::{CommandContext, ControlCommandHandler};
 pub use help::Help;
 pub use list::List;
 pub use mdtm::Mdtm;
