+++ src/server/controlchan/handlers/nlst.rs
@@ -14,8 +14,8 @@
 // the implementation of a "multiple get" function.
 
 use super::handler::CommandContext;
+use crate::server::controlchan::command::Command;
 use crate::server::controlchan::handlers::ControlCommandHandler;
-use crate::server::controlchan::handlers::Command;
 use crate::server::error::FTPError;
 use crate::server::reply::{Reply, ReplyCode};
 use crate::storage;
